+++ src/main/kotlin/kr/hhplus/be/server/service/product/usecase/AddProductStockUsecase.kt
@@ -6,5 +6,5 @@ interface AddProductStockUsecase {
     /**
      * 상품 재고 증가 처리
      */
-    fun addProductStock(productId: Long, quantity: Int, now: ZonedDateTime)
+    fun addProductStock(productId: Long, quantity: Long, now: ZonedDateTime)
 }
